Move artifacts out of src folder
