+++ yarn-project/.gitignore
@@ -12,10 +12,9 @@
 **/.debounce-*
 **/.tsc.pid
 **/*.result
-accounts/src/artifacts
+accounts/artifacts
 aztec-faucet/data*
 aztec-node/data*
-aztec-js/src/account_contract/artifacts
 aztec/log
 circuits.js/fixtures/*.json
 circuits.js/src/structs/kernel/private_kernel_reset_circuit_private_inputs_variants.ts
@@ -25,7 +24,6 @@ end-to-end/log
 end-to-end/data
 end-to-end/src/web/main.js
 end-to-end/src/web/main.js.LICENSE.txt
-entry-points/src/artifacts
 l1-artifacts/generated
 l1-contracts/generated
 builder/target/
@@ -40,7 +38,7 @@ builder/src/target
 builder/src/crs
 builder/src/types
 noir-protocol-circuits-types/src/types/
-protocol-contracts/src/artifacts
+protocol-contracts/artifacts
 scripts/tmp
 noir-contracts.js/src
 noir-contracts.js/artifacts/
