+++ yarn-project/accounts/.prettierignore
@@ -1,1 +1,1 @@
-src/artifacts/*.json
+artifacts/*.json
