+++ yarn-project/accounts/src/ecdsa/artifact.ts
@@ -1,5 +1,5 @@
 import { type NoirCompiledContract, loadContractArtifact } from '@aztec/aztec.js';
 
-import EcdsaAccountContractJson from '../artifacts/EcdsaAccount.json' assert { type: 'json' };
+import EcdsaAccountContractJson from '../../artifacts/EcdsaAccount.json' assert { type: 'json' };
 
 export const EcdsaAccountContractArtifact = loadContractArtifact(EcdsaAccountContractJson as NoirCompiledContract);
