+++ yarn-project/accounts/src/schnorr/artifact.ts
@@ -1,5 +1,5 @@
 import { type NoirCompiledContract, loadContractArtifact } from '@aztec/aztec.js';
 
-import SchnorrAccountContractJson from '../artifacts/SchnorrAccount.json' assert { type: 'json' };
+import SchnorrAccountContractJson from '../../artifacts/SchnorrAccount.json' assert { type: 'json' };
 
 export const SchnorrAccountContractArtifact = loadContractArtifact(SchnorrAccountContractJson as NoirCompiledContract);
