+++ yarn-project/accounts/src/single_key/artifact.ts
@@ -1,6 +1,6 @@
 import { type NoirCompiledContract, loadContractArtifact } from '@aztec/aztec.js';
 
-import SchnorrSingleKeyAccountContractJson from '../artifacts/SchnorrSingleKeyAccount.json' assert { type: 'json' };
+import SchnorrSingleKeyAccountContractJson from '../../artifacts/SchnorrSingleKeyAccount.json' assert { type: 'json' };
 
 export const SchnorrSingleKeyAccountContractArtifact = loadContractArtifact(
   SchnorrSingleKeyAccountContractJson as NoirCompiledContract,
