+++ yarn-project/protocol-contracts/src/auth-registry/artifact.ts
@@ -1,6 +1,6 @@
 import { loadContractArtifact } from '@aztec/types/abi';
 import { type NoirCompiledContract } from '@aztec/types/noir';
 
-import AuthRegistryJson from '../artifacts/AuthRegistry.json' assert { type: 'json' };
+import AuthRegistryJson from '../../artifacts/AuthRegistry.json' assert { type: 'json' };
 
 export const AuthRegistryArtifact = loadContractArtifact(AuthRegistryJson as NoirCompiledContract);
