+++ yarn-project/protocol-contracts/src/class-registerer/artifact.ts
@@ -1,7 +1,7 @@
 import { loadContractArtifact } from '@aztec/types/abi';
 import { type NoirCompiledContract } from '@aztec/types/noir';
 
-import ContractClassRegistererJson from '../artifacts/ContractClassRegisterer.json' assert { type: 'json' };
+import ContractClassRegistererJson from '../../artifacts/ContractClassRegisterer.json' assert { type: 'json' };
 
 export const ContractClassRegistererArtifact = loadContractArtifact(
   ContractClassRegistererJson as NoirCompiledContract,
