+++ yarn-project/protocol-contracts/src/instance-deployer/artifact.ts
@@ -1,7 +1,7 @@
 import { loadContractArtifact } from '@aztec/types/abi';
 import { type NoirCompiledContract } from '@aztec/types/noir';
 
-import ContractInstanceDeployerJson from '../artifacts/ContractInstanceDeployer.json' assert { type: 'json' };
+import ContractInstanceDeployerJson from '../../artifacts/ContractInstanceDeployer.json' assert { type: 'json' };
 
 export const ContractInstanceDeployerArtifact = loadContractArtifact(
   ContractInstanceDeployerJson as NoirCompiledContract,
