+++ yarn-project/protocol-contracts/src/multi-call-entrypoint/artifact.ts
@@ -1,6 +1,6 @@
 import { loadContractArtifact } from '@aztec/types/abi';
 import { type NoirCompiledContract } from '@aztec/types/noir';
 
-import MultiCallEntrypoint from '../artifacts/MultiCallEntrypoint.json' assert { type: 'json' };
+import MultiCallEntrypoint from '../../artifacts/MultiCallEntrypoint.json' assert { type: 'json' };
 
 export const MultiCallEntrypointArtifact = loadContractArtifact(MultiCallEntrypoint as NoirCompiledContract);
