+++ yarn-project/protocol-contracts/tsconfig.json
@@ -16,5 +16,5 @@
       "path": "../types"
     }
   ],
-  "include": ["src", "src/artifacts/*.d.json.ts"]
+  "include": ["src", "artifacts/*.d.json.ts"]
 }
